Refactor index.py into various classes/files Most of the app is broken now
